Use one centralised InterleavePending type for futures-test
