+++ futures-test/src/io/write/mod.rs
@@ -2,8 +2,8 @@
 
 use futures_io::AsyncWrite;
 
-pub use super::interleave_pending::InterleavePending;
 pub use super::limited::Limited;
+pub use crate::interleave_pending::InterleavePending;
 
 /// Additional combinators for testing async writers.
 pub trait AsyncWriteTestExt: AsyncWrite {
